Replaced CDN JS with NPM/local
