+++ _config/filepond.yml
@@ -1,15 +1,23 @@
 ---
 Name: filepond
 ---
+SilverStripe\Admin\LeftAndMain:
+  extra_requirements_javascript:
+    - 'lekoala/silverstripe-filepond:client/dist/js/FilePondField.js'
+  extra_requirements_css:
+    - 'lekoala/silverstripe-filepond:client/dist/styles/FilePondField.css'
+
 LeKoala\FilePond\FilePondField:
   auto_clear_temp_folder: false
   auto_clear_threshold: null #defaults to 1 day in prod env
   enable_default_description: true
   enable_requirements: true
+
 SilverStripe\Assets\File:
   enable_auto_thumbnails: false
   extensions:
     - LeKoala\FilePond\FilePondFileExtension
+
 # ---
 # Name: filepond-simpler
 # After: "#simpler_silverstripe"
