+++ client/src/styles/FilePondField.scss
@@ -1,0 +1,3 @@
+@import '~filepond/dist/filepond.css';
+@import '~filepond-plugin-file-poster/dist/filepond-plugin-file-poster.css';
+@import '~filepond-plugin-image-preview/dist/filepond-plugin-image-preview.css';
